refactor: pull watch frontend into separate file
